BiBlock

A composite unit that represents a block. It can be used for sampling on a block in a setting of smoothing or inference.

DiffusionMCMCTools.BiBlock — Type

mutable struct BiBlock{L,TGP,TGPl,TW,TWn,TX}
    b::Block{L,TGP,TGPl,TW,TWn,TX}
    b°::Block{L,TGP,TGPl,TW,TWn,TX}
    ρ::Float64
    accpt_history::Vector{Bool}
end

Composite unit that allows for sampling of a single block. It provides two Blocks: one proposal b°, one accepted b that can be used for smoothing or inference problems. ρ is a memory parameter of the preconditioned Crank-Nicolson scheme and accpt_history stores the history of accept/reject decisions (useful for MCMC).

function BiBlock(
    sp::SamplingPair,
    range::UnitRange{Int64},
    ρ=0.0,
    last_block=false,
    ll_hist_len=0
)

Base constructor.

There are many functions implemented for it.

For adjustments made after the accept-reject decision (regardless of what it was)

DiffusionMCMCTools.set_accepted! — Method

set_accepted!(bb::BiBlock, i::Int, v)

Commit the accept/reject decision v to acceptance history of BiBlock b at the position i.

For adjustments to the containers in case of acceptance of proposals:

DiffusionMCMCTools.swap_paths! — Method

swap_paths!(bb::BiBlock)

Swap XX and WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_XX! — Method

swap_XX!(bb::BiBlock)

Swap XX containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_WW! — Method

swap_WW!(bb::BiBlock)

Swap WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_PP! — Method

swap_PP!(bb::BiBlock)

Swap PP containers (including PP_last) between proposal-acceptance pair.