Block

A smallest unit that is needed for blocking.

DiffusionMCMCTools.Block — Type

mutable struct Block{L,TGP,TGPl,TW,TWn,TX}
    PP::TVIEW{TGP}
    P_last::TVIEW{TGPl} # view into a single element
    WW::TVIEW{TW}
    Wnr::TWn
    XX::TVIEW{TX}
    ll::Float64
    ll_history::Vector{Float64}
end

The smallest "containerless" unit that provides a view into SamplingUnit restricted to a range i:j of a block. L is an important flag that indicates whether it is a terminal block or not.

Fields

PP: a vector of views into relevant GuidProp

P_last: a view into a single (or none) GuidProp corresponding to the terminal sub-interval.

WW: a vector of views into containers for sampled Wiener process

Wnr: a flag for sampling Wiener processes

XX: a vector of views into containers for a sampled process

ll: a placeholder for computed log-likelihood

ll_history: a history of computed log-likelihoods (useful for MCMC)

function Block(
    u::SamplingUnit,
    range::UnitRange{Int64},
    last_block=false,
    ll_hist_len=0
)

Base constructor.

It is rarely used on its own. Instead, it is used mainly as a building block of other, composite units. Nevertheless, there are a couple of useful functions implemented for it:

DiffusionMCMCTools.set_ll! — Method

set_ll!(b::Block, i::Int, v)

Set the internal log-likelihood history field b.ll_history[i] with a value v

DiffusionMCMCTools.save_ll! — Method

save_ll!(b::Block, i::Int)

Commit the current log-likelihood field ll to history b.ll_history at index i.

GuidedProposals.recompute_guiding_term! — Method

GP.recompute_guiding_term!(b::Block)

Recompute the guiding term.

DiffusionMCMCTools.find_W_for_X! — Method

find_W_for_X!(b::Block)

Compute the Wiener process b.WW that is needed for obtaining path b.XX under the law stored in b.

GuidedProposals.loglikhd — Method

GP.loglikhd(b::Block)

Compute the log-likelihood evaluated at a sampled path.

DiffusionMCMCTools.loglikhd! — Method

loglikhd!(b::Block)

Compute the log-likelihood evaluated at a sampled path and store the result in an internal field ll.

DiffusionMCMCTools.recompute_path! — Function

recompute_path!(b::Block, WW=b.WW; skip=0)

Recompute the path b.XX for a given wiener process WW.