BlockCollection

DiffusionMCMCTools.BlockCollection — Type

struct BlockCollection{TGP,TGPl,TW,TWn,TX}
    blocks::Vector{BiBlock{_L,TGP,TGPl,TW,TWn,TX} where _L}
end

Gathers all blocks relevant for an entire single recording.

BlockCollection(
    sp::SamplingPair{TGP,TGPl,TW,TWn,TX},
    ranges,
    ρρ=0.0,
    ll_hist_len=0
) where {TGP,TGPl,TW,TWn,TX}

Base constructor.

Imputation

DiffusionMCMCTools.draw_proposal_path! — Method

draw_proposal_path!(bc::BlockCollection)

Sample proposal paths on each block, compute log-likelihoods along the way. Assumes bc.blocks[i].b.XX[1].x[1] are starting points for each block. Uses the preconditioned Crank-Nicolson scheme.

Accept/reject decision

DiffusionMCMCTools.accept_reject_proposal_path! — Method

accept_reject_proposal_path!(bb::BlockCollection, mcmciter)

Accept/reject decision of the Metropolis-Hastings algorithm for the step of path imputation, done separately for each block.

Swaps

DiffusionMCMCTools.swap_paths! — Method

swap_paths!(bc::BlockCollection)

For each block in the collection swap XX and WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_XX! — Method

swap_XX!(bc::BlockCollection)

For each block in the collection swap XX containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_WW! — Method

swap_WW!(bc::BlockCollection)

For each block in the collection swap WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_PP! — Method

swap_PP!(bc::BlockCollection)

For each block in the collection swap PP containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_ll! — Method

swap_ll!(bc::BlockCollection)

For each block in the collection swap ll fields between proposal-acceptance pair.

Setting up a block

GuidedProposals.set_obs! — Method

GP.set_obs!(bc::BlockCollection)

Freeze an artificial observation at the terminal point of each block. For a terminal block nothing is done.

utility

DiffusionMCMCTools.loglikhd! — Method

loglikhd!(bc::BlockCollection)

For each BiBlock in a collection compute the log-likelihood for the accepted block, evaluated at a sampled path and store the result in internal fields ll.

DiffusionMCMCTools.loglikhd°! — Method

loglikhd°!(bc::BlockCollection)

For each BiBlock in a collection compute the log-likelihood for the proposal block, evaluated at a sampled path and store the result in internal fields ll.

DiffusionMCMCTools.fetch_ll — Method

fetch_ll(bc::BlockCollection)

Retreive the log-likelihood for the entire accepted path.

Warning

The function uses only internal fields ll for this computation, which means that for the call to this function to make sense the log-likelihood must have been previously computed and stored in the field ll. If it hasn't been done, then you must first call loglikhd!(bc).

DiffusionMCMCTools.fetch_ll° — Method

fetch_ll°(bc::BlockCollection)

Retreive the log-likelihood for the entire proposed path.

Warning

The function uses only internal fields ll for this computation, which means that for the call to this function to make sense the log-likelihood must have been previously computed and stored in the field ll. If it hasn't been done, then you must first call loglikhd°!(bc).

DiffusionMCMCTools.save_ll! — Method

save_ll!(bc::BlockCollection, i::Int)

For each block in the collection commit the current proposal and accepted log-likelihood fields ll to history, at index i.

DiffusionMCMCTools.ll_of_accepted — Method

ll_of_accepted(bb::BlockCollection, i)

Return an array of log-likelihoods (one for each block) of the paths that were accepted at the ith iteration.

DiffusionMCMCTools.accpt_rate — Method

accpt_rate(bb::BlockCollection, range)

Compute the acceptance rate over the range of MCMC accept/reject history for each block in the collection.

Setting parameters

DiffusionMCMCTools.set_proposal_law! — Function

set_proposal_law!(
    bc::BlockCollection,
    θ°,
    pnames,
    critical_change=GP.is_critical_update(bc, pnames);
    skip=0
)

For each block in the collection set the parameters in bb.b°.PP and bb.b°.P_last to θ° and make sure that all other parameters are shared with bb.b.PP and bb.b.P_last. Recompute the guiding term if needed, and then, compute the proposal trajectory bb.b°.XX for the proposal point θ°.

Example

The algorithm

all_obs = AllObservations()
add_recording!(all_obs, build_recording(P, data, 0.0, KnownStartingPt(y1)))
DD.var_parameter_names(::FitzHughNagumo) = (:γ,)
all_obs, _ = initialize(all_obs)

customkernel(θ, scale=0.1) = θ .+ 2.0*scale*(rand()-0.5)

function accept_reject_proposal_param!(bc, mcmciter, θ, θ°)
    accepted = rand(Exponential(1.0)) > -(fetch_ll°(bc)-fetch_ll(bc))
    accepted && swap_XX!(bc)
    accepted && swap_PP!(bc)
    save_ll!(bc, mcmciter)
    accepted && swap_ll!(bc)
    accepted, copy(accepted ? θ° : θ)
end

function simple_inference(AuxLaw, all_obs, dt, _θ; ϵ=0.3, ρ=0.5, num_steps=10^4)
    # making sure that things are in order...
    _pname = collect(keys(_θ))
    # for simplicity restrict to inference for a single param
    @assert length(_pname) == 1
    pname = first(_pname)
    θ = collect(values(_θ))

    # setting the initial guess θ inside the recording
    OBS.set_parameters!(all_obs, _θ)
    @assert num_recordings(all_obs) == 1
    recording = first(all_obs.recordings)

    # setting up containers
    num_obs = length(recording.obs)
    tts = OBS.setup_time_grids(recording, dt, standard_guid_prop_time_transf)
    sp = SamplingPair(AuxLaw, recording, tts)
    bc = BlockCollection(sp, [1:num_obs], ρ, num_steps)
    name_struct = ParamNamesRecording(
        bc, _pname, first(all_obs.param_depend_rev),
        first(all_obs.obs_depend_rev)
    )

    loglikhd!(bc)
    paths = []

    θθ = [θ]
    a_h = Bool[]
    crit_change = [true]

    for i in 1:num_steps
        draw_proposal_path!(bc)
        accept_reject_proposal_path!(bc, i)

        θ° = customkernel(θ, ϵ)
        set_proposal_law!(bc, θ°, name_struct, crit_change)

        accpt, θ = accept_reject_proposal_param!(bc, i, θ, θ°)
        push!(θθ, θ)
        push!(a_h, accpt)

        # progress message
        if i % 100 == 0
            println(
                "$i. ll=$(ll_of_accepted(bc, i)), ",
                "imp a-r: ",
                " $(accpt_rate(bc, (i-99):i)), ",
                "updt a-r: ",
                "$(sum(a_h[(i-99):i])/100)."
            )
        end

        # save intermediate path for plotting
        i % 400 == 0 && append!(paths, [deepcopy(sp.u.XX)])
    end
    paths, θθ
end

Result

using OrderedCollections

θ = OrderedDict(:REC1_γ=>1.5)

DD.var_parameter_names(::FitzHughNagumo) = (:γ,)
DD.var_parameter_names(::FitzHughNagumoAux) = (:γ,)

@load_diffusion FitzHughNagumoAux
paths, θθ = simple_inference(
    FitzHughNagumoAux, all_obs, 0.001, θ; ϵ=0.3, ρ=0.96, num_steps=10^4
)

Example of inference with blocking

function accept_reject_proposal_param!(bc::BlockCollection, mcmciter, θ, θ°)
    accepted = rand(Exponential(1.0)) > -(fetch_ll°(bc) - fetch_ll(bc))
    accepted && swap_XX!(bc)
    accepted && swap_PP!(bc)
    save_ll!(bc, mcmciter)
    accepted && swap_ll!(bc)
    accepted, copy(accepted ? θ° : θ)
end

function simple_inference_with_blocking(
        AuxLaw, all_obs, dt, AuxLawBlocking, block_layout, _θ;
        ϵ=0.3, ρ=0.5, num_steps=10^4
    )
    # making sure that things are in order...
    _pname = collect(keys(_θ))
    # for simplicity restrict to inference for a single param
    @assert length(_pname) == 1
    pname = first(_pname)
    θ = collect(values(_θ))

    # setting the initial guess θ inside the recording
    OBS.set_parameters!(all_obs, _θ)
    @assert num_recordings(all_obs) == 1
    recording = first(all_obs.recordings)

    # setting up containers
    tts = OBS.setup_time_grids(recording, dt, standard_guid_prop_time_transf)
    sp = SamplingPair(AuxLaw, recording, tts)
    blocks = [
        BlockCollection(sp, block_ranges, ρ, num_steps)
        for block_ranges in block_layout
    ]
    name_struct = [
        ParamNamesRecording(
            bc, _pname, first(all_obs.param_depend_rev),
            first(all_obs.obs_depend_rev)
        ) for bc in blocks
    ]

    paths = []

    θθ = [θ]
    a_h = Bool[]
    crit_change = [fill(true, length(bc.blocks)) for bc in blocks]

    # MCMC
    for i in 1:num_steps
        for bc in blocks
            GP.set_obs!(bc)
            recompute_guiding_term!(bc, Val(:P_only))
            find_W_for_X!(bc)
            loglikhd!(bc)
            draw_proposal_path!(bc)
            accept_reject_proposal_path!(bc, i)

            # progress message
            if i % 100 == 0
                println(
                    "$i. ll=$(ll_of_accepted(bc, i)), acceptance rate: ",
                    "$( accpt_rate(bc, (i-99):i) )"
                )
            end
        end

        θ° = customkernel(θ, ϵ)

        bc = blocks[end]
        set_proposal_law!(bc, θ°, name_struct[end], crit_change[end])
        recompute_guiding_term!(bc, Val(:P°_only))

        accpt, θ = accept_reject_proposal_param!(bc, i, θ, θ°)
        push!(θθ, θ)
        push!(a_h, accpt)

        if i % 100 == 0
            println(
                "$i. updt a-r: ",
                "$(sum(a_h[(i-99):i])/100)."
            )
        end

        # save intermediate path for plotting
        i % 400 == 0 && append!(paths, [deepcopy(sp.u.XX)])
    end
    paths, θθ
end