Containers for parameter names

DiffusionMCMCTools.ParamNamesUnit — Type

struct ParamNamesUnit{N1,N2}
    var::NTuple{N1,Symbol}
    var_aux::Vector{Tuple{Vararg{Symbol,_N} where _N}}
    updt::NTuple{N2,Pair{Int64,Symbol}}
    updt_aux::Vector{Tuple{Vararg{Pair{Int64,Symbol},_N} where _N}}
    updt_obs::Vector{Tuple{Vararg{Pair{Int64,Int64},_N} where _N}}
end

Smallest unit storing information about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It is pertinent to a single block of a single recording only, and in particular, only to a single collection of laws, out of:

PP, i.e. those relevant to a given block

P_last, i.e. the artificial law of the block associated with the last obs

P_excl, i.e. the last, missed out PP, missed out due to use of P_last

PPb, i.e. the collection of all remaining artificial laws that are not P_last

It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

Fields

var: names of P_target that are checked and—if need be—equalized between proposal and accepted laws

var_aux: names of P_aux that are checked and—if need be—equalized between proposal and accepted laws

updt: list of params of P_target that are being updated at a given MCMC step; listed as pairs: idx-of-θ-to-relevant-value => name-of-param-inside-P_target-struct

updt_aux: as above, but for P_aux

updt_obs: list of params of obs (i.e. terminal observation for each guid prop) that are being updated at a given MCMC step; listed as pairs: idx-of-θ-to-relevant-value => idx-of-obs-inside-obs.θ

ParamNamesUnit(
    PP::AbstractArray{<:GuidProp}, θnames::Vector{Symbol}, pdep, odeps
)

Base constructor.

Arguments

PP:

θnames:

pdep:

odeps:

DiffusionMCMCTools.ParamNamesBlock — Type

struct ParamNamesBlock{N1,N2}
    PP::ParamNamesUnit{N1,N2}
    P_last::ParamNamesUnit{N1,N2}
    P_excl::ParamNamesUnit{N1,N2}
    Pb_excl::ParamNamesUnit{N1,N2}
end

Stores information relevant for a Block about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

Fields

PP, i.e. those relevant to a given block

P_last, i.e. the artificial law of the block associated with the last obs

P_excl, i.e. the last, missed out PP, missed out due to use of P_last

Pb_excl, i.e. the collection of all remaining artificial laws that are not P_last

ParamNamesBlock(b::Block, θnames, pdep, odeps)

Base constructor.

Arguments

b::Block:

θnames:

pdep:

odeps:

DiffusionMCMCTools.ParamNamesRecording — Type

struct ParamNamesRecording{N1,N2}
    blocks::Vector{ParamNamesBlock{N1,N2}}
end

Stores information relevant for an entire, single recording about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

ParamNamesRecording(rb::BlockCollection, θnames, pdep, odeps)

Base constructor.

DiffusionMCMCTools.ParamNamesAllObs — Type

struct ParamNamesAllObs{T}
    recordings::Vector{T}
end

Stores information relevant for multiple recordings about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

ParamNamesAllObs(be::BlockEnsemble, θnames, all_obs)

Base constructor.