A pairing of two SamplingUnits

Defines all main containers for an entire single recording for a smoothing or inference problem.

DiffusionMCMCTools.SamplingPair — Type

struct SamplingPair{TGP,TGPb,TW,TWn,TX}
    u::SamplingUnit{TGP,TGPb,TW,TWn,TX}
    u°::SamplingUnit{TGP,TGPb,TW,TWn,TX}
end

A pairing of two SamplingUnits, can be used for smoothing or inference problems.

SamplingPair(
    aux_laws, recording, tts, args=tuple();
    aux_laws_blocking=aux_laws, artificial_noise=1e-11,
    solver_choice_blocking=args
)

Base constructor.

Arguments

aux_laws:

recording:

tts:

args:

aux_laws_blocking:

artificial_noise:

solver_choice_blocking:

Note

In practice all sampling is done with a BiBlock that looks at sections of a SamplingPair and never through SamplingPair directly. Even when no blocking is needed, it should still be done by defining a BiBlock that simply looks at the entire SamplingPair.