BiBlock

A composite unit that represents a block. It can be used for sampling on a block in a setting of smoothing or inference.

DiffusionMCMCTools.BiBlock — Type

mutable struct BiBlock{L,TGP,TGPl,TW,TWn,TX}
    b::Block{L,TGP,TGPl,TW,TWn,TX}
    b°::Block{L,TGP,TGPl,TW,TWn,TX}
    ρ::Float64
    accpt_history::Vector{Bool}
end

Composite unit that allows for sampling of a single block. It provides two Blocks: one proposal b°, one accepted b that can be used for smoothing or inference problems. ρ is a memory parameter of the preconditioned Crank-Nicolson scheme and accpt_history stores the history of accept/reject decisions (useful for MCMC).

function BiBlock(
    sp::SamplingPair,
    range::UnitRange{Int64},
    ρ=0.0,
    last_block=false,
    ll_hist_len=0
)

Base constructor.

There are many functions implemented for it.

Imputation of paths

DiffusionMCMCTools.draw_proposal_path! — Method

draw_proposal_path!(bb::BiBlock)

Sample a proposal path, compute log-likelihood along the way. Assumes bb.b.XX[1].x[1] is a starting point. Uses preconditioned Crank-Nicolson scheme with memory parameter set as bb.ρ.

Accept/reject decision in an MCMC setting

DiffusionMCMCTools.accept_reject_proposal_path! — Method

accept_reject_proposal_path!(bb::BiBlock, mcmciter)

Accept/reject decision of the Metropolis-Hastings algorithm for the step of path imputation.

Adjustments made after the accept-reject decision (regardless of what it was)

DiffusionMCMCTools.set_accepted! — Method

set_accepted!(bb::BiBlock, i::Int, v)

Commit the accept/reject decision v to acceptance history of BiBlock b at the position i.

Adjustments to the containers in case of acceptance of proposals:

DiffusionMCMCTools.swap_paths! — Method

swap_paths!(bb::BiBlock)

Swap XX and WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_XX! — Method

swap_XX!(bb::BiBlock)

Swap XX containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_WW! — Method

swap_WW!(bb::BiBlock)

Swap WW containers between proposal-acceptance pair.

DiffusionMCMCTools.swap_PP! — Method

swap_PP!(bb::BiBlock)

Swap PP containers (including PP_last) between proposal-acceptance pair.

Setting up a block

GuidedProposals.set_obs! — Method

GP.set_obs!(bb::BiBlock)

Freeze an artificial observation at the terminal point of the block. For a terminal block nothing is done.

GuidedProposals.recompute_guiding_term! — Method

GP.recompute_guiding_term!(bb::BiBlock)

Recompute the guiding terms of both the proposal and the accepted laws.

DiffusionMCMCTools.find_W_for_X! — Method

find_W_for_X!(bb::BiBlock)

Find the Wiener process bb.b.WW that reconstructs path bb.b.XX under the accepted law bb.b.PP (possibly including bb.b.P_last).

Utility

DiffusionMCMCTools.ll_of_accepted — Method

ll_of_accepted(bb::BiBlock, i)

Return the log-likelihood of the path that was accepted at the ith iteration.