Containers for parameter names

DiffusionMCMCTools.ParamNamesUnit — Type

struct ParamNamesUnit{N1,N2}
    var::NTuple{N1,Symbol}
    var_aux::Vector{Tuple{Vararg{Symbol,_N} where _N}}
    updt::NTuple{N2,Pair{Int64,Symbol}}
    updt_aux::Vector{Tuple{Vararg{Pair{Int64,Symbol},_N} where _N}}
    updt_obs::Vector{Tuple{Vararg{Pair{Int64,Int64},_N} where _N}}
end

Smallest unit storing information about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It is pertinent to a single block of a single recording only, and in particular, only to a single collection of laws, out of:

PP, i.e. those relevant to a given block

P_last, i.e. the artificial law of the block associated with the last obs

P_excl, i.e. the last, missed out PP, missed out due to use of P_last

PPb, i.e. the collection of all remaining artificial laws that are not P_last

It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

Fields

var: names of P_target that are checked and—if need be—equalized between proposal and accepted laws

var_aux: names of P_aux that are checked and—if need be—equalized between proposal and accepted laws

updt: list of params of P_target that are being updated at a given MCMC step; listed as pairs: idx-of-θ-to-relevant-value => name-of-param-inside-P_target-struct

updt_aux: as above, but for P_aux

updt_obs: list of params of obs (i.e. terminal observation for each guid prop) that are being updated at a given MCMC step; listed as pairs: idx-of-θ-to-relevant-value => idx-of-obs-inside-obs.θ

ParamNamesUnit(
    PP::AbstractArray{<:GuidProp}, θnames::Vector{Symbol}, pdep, odeps
)

Base constructor.

Arguments

PP:

θnames:

pdep:

odeps:

DiffusionMCMCTools.ParamNamesBlock — Type

struct ParamNamesBlock{N1,N2}
    PP::ParamNamesUnit{N1,N2}
    P_last::ParamNamesUnit{N1,N2}
    P_excl::ParamNamesUnit{N1,N2}
    Pb_excl::ParamNamesUnit{N1,N2}
end

Stores information relevant for a Block about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

Fields

PP, i.e. those relevant to a given block

P_last, i.e. the artificial law of the block associated with the last obs

P_excl, i.e. the last, missed out PP, missed out due to use of P_last

Pb_excl, i.e. the collection of all remaining artificial laws that are not P_last

ParamNamesBlock(b::Block, θnames, pdep, odeps)

Base constructor.

Arguments

b::Block:

θnames:

pdep:

odeps:

DiffusionMCMCTools.ParamNamesRecording — Type

struct ParamNamesRecording{N1,N2}
    blocks::Vector{ParamNamesBlock{N1,N2}}
end

Stores information relevant for an entire, single recording about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

ParamNamesRecording(rb::BlockCollection, θnames, pdep, odeps)

Base constructor.

DiffusionMCMCTools.ParamNamesAllObs — Type

struct ParamNamesAllObs{T}
    recordings::Vector{T}
end

Stores information relevant for multiple recordings about the names of the parameters that are supposed to be updated at a given step of the MCMC algorithm. It stores lists of parameters that are supposed to be updated at various stages of the call to a function set_proposal_law!, which sets the proposal parameter θ° inside the proposal laws.

ParamNamesAllObs(be::BlockEnsemble)

Base constructor.

Example

The algorithm

all_obs = AllObservations()
add_recording!(all_obs, build_recording(P, data, 0.0, KnownStartingPt(y1)))
DD.var_parameter_names(::FitzHughNagumo) = (:γ,)
all_obs, _ = initialize(all_obs)

customkernel(θ, scale=0.1) = θ .+ 2.0*scale*(rand()-0.5)

#=
    NOTE we no longer need to hand-craft the `_build_struct` and
    `simple_name_structure` functions. These will be substituted with a
    struct `ParamNamesRecording`
=#

function accept_reject_proposal_param!(bb, mcmciter, θ, θ°)
    accepted = rand(Exponential(1.0)) > -(bb.b°.ll - bb.b.ll)
    accepted && swap_XX!(bb)
    accepted && swap_PP!(bb)
    save_ll!(bb.b, mcmciter)
    save_ll!(bb.b°, mcmciter)
    accepted && swap_ll!(bb)
    accepted, copy(accepted ? θ° : θ)
end

function simple_inference(AuxLaw, all_obs, dt, _θ; ϵ=0.3, ρ=0.5, num_steps=10^4)
    # making sure that things are in order...
    _pname = collect(keys(_θ))
    # for simplicity restrict to inference for a single param
    @assert length(_pname) == 1
    pname = first(_pname)
    θ = collect(values(_θ))

    # setting the initial guess θ inside the recording
    OBS.set_parameters!(all_obs, _θ)
    @assert num_recordings(all_obs) == 1
    recording = first(all_obs.recordings)

    # setting up containers
    num_obs = length(recording.obs)
    tts = OBS.setup_time_grids(recording, dt, standard_guid_prop_time_transf)
    sp = SamplingPair(AuxLaw, recording, tts)
    bb = BiBlock(sp, 1:num_obs, ρ, true, num_steps)
    name_struct = ParamNamesBlock(
        bb.b, _pname, first(all_obs.param_depend_rev),
        first(all_obs.obs_depend_rev)
    )

    loglikhd!(bb)
    paths = []

    θθ = [θ]
    a_h = Bool[]

    for i in 1:num_steps
        draw_proposal_path!(bb)
        accept_reject_proposal_path!(bb, i)

        θ° = customkernel(θ, ϵ)
        set_proposal_law!(bb, θ°, name_struct, true)

        accpt, θ = accept_reject_proposal_param!(bb, i, θ, θ°)
        push!(θθ, θ)
        push!(a_h, accpt)

        # progress message
        if i % 100 == 0
            println(
                "$i. ll=$(ll_of_accepted(bb, i)), ",
                "imp a-r: ",
                " $(accpt_rate(bb, (i-99):i)), ",
                "updt a-r: ",
                "$(sum(a_h[(i-99):i])/100)."
            )
        end

        # save intermediate path for plotting
        i % 400 == 0 && append!(paths, [deepcopy(sp.u.XX)])
    end
    paths, θθ
end

Result

using OrderedCollections

θ = OrderedDict(:REC1_γ=>1.5)

DD.var_parameter_names(::FitzHughNagumo) = (:γ,)
DD.var_parameter_names(::FitzHughNagumoAux) = (:γ,)

@load_diffusion FitzHughNagumoAux
paths, θθ = simple_inference(
    FitzHughNagumoAux, all_obs, 0.001, θ; ϵ=0.3, ρ=0.96, num_steps=10^4
)